feat: add modal for creating new content and update content API to handle optional fields

+++ server/api/content/update.post.ts
@@ -3,11 +3,11 @@ import { z } from "zod";
 const bodySchema = z.object({
   id: z.string(),
   title: z.string(),
-  description: z.string(),
-  date: z.string().nullable(),
-  url: z.string(),
-  image: z.string().nullable(),
-  categories: z.array(z.number()),
+  description: z.string().nullable().optional(),
+  date: z.string().nullable().optional(),
+  url: z.string().nullable().optional(),
+  image: z.string().nullable().optional(),
+  categories: z.array(z.number()).nullable().optional(),
 })
 
 export default defineEventHandler(async (event) => {
@@ -17,6 +17,9 @@ export default defineEventHandler(async (event) => {
   const { id, categories, title, description, date, url, image } = await readValidatedBody(event, body => bodySchema.parse(body))
 
   await db.updateTable('contents').set({ title, description, date, url, image }).where('id', '=', Number(id)).execute()
-  await db.deleteFrom('contentCategories').where('contentId', '=', Number(id)).execute()
-  await db.insertInto('contentCategories').values(categories.map(categoryId => ({ contentId: Number(id), categoryId }))).execute()
+
+  if (categories && categories.length > 0) {
+    await db.deleteFrom('contentCategories').where('contentId', '=', Number(id)).execute()
+    await db.insertInto('contentCategories').values(categories.map(categoryId => ({ contentId: Number(id), categoryId }))).execute()
+  }
 })
